Time statistics adjusted, now it's right. Async segments created on IO and cloned on Clusters. Testing PUT/GET Op via Cluster to Cluster (SUCCESSFULL) and Cluster To IO (Failed)

+++ mppa256/async/include/timer.h
@@ -16,7 +16,7 @@
 /*
  * Computers the difference between two times
  */
-extern uint64_t timer_diff(uint64_t t1, uint64_t t2);
+extern float timer_diff(uint64_t t1, uint64_t t2);
 
 /*
  * Gets the current timer value
